+++ .github/dependabot.yml
@@ -29,15 +29,6 @@ updates:
       interval: weekly
       day: sunday
   - package-ecosystem: gomod
-    directory: /bridge/opencensus/opencensusmetric
-    labels:
-      - dependencies
-      - go
-      - Skip Changelog
-    schedule:
-      interval: weekly
-      day: sunday
-  - package-ecosystem: gomod
     directory: /bridge/opencensus/test
     labels:
       - dependencies
