+++ bridge/opencensus/test/go.mod
@@ -14,9 +14,8 @@ require (
 	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
-	go.opentelemetry.io/otel/bridge/opencensus/opencensusmetric v0.0.0-00010101000000-000000000000 // indirect
-	go.opentelemetry.io/otel/metric v0.0.0-00010101000000-000000000000 // indirect
-	go.opentelemetry.io/otel/sdk/metric v0.0.0-00010101000000-000000000000 // indirect
+	go.opentelemetry.io/otel/metric v0.32.0 // indirect
+	go.opentelemetry.io/otel/sdk/metric v0.32.0 // indirect
 	golang.org/x/sys v0.0.0-20210423185535-09eb48e85fd7 // indirect
 )
 
@@ -31,5 +30,3 @@ replace go.opentelemetry.io/otel/trace => ../../../trace
 replace go.opentelemetry.io/otel/metric => ../../../metric
 
 replace go.opentelemetry.io/otel/sdk/metric => ../../../sdk/metric
-
-replace go.opentelemetry.io/otel/bridge/opencensus/opencensusmetric => ../opencensusmetric
